게시물 삭제시 권한체크

+++ src/main/java/com/wej/exam/demo/controller/UsrArticleController.java
@@ -26,16 +26,16 @@ public class UsrArticleController {
 
 		boolean isLogined = false;
 		int loginedMemberId = 0;
-		
-		if( httpSession.getAttribute("loginedMemberId") != null ) {
+
+		if (httpSession.getAttribute("loginedMemberId") != null) {
 			isLogined = true;
 			loginedMemberId = (int) httpSession.getAttribute("loginedMemberId");
 		}
-		
-		if (isLogined == false ) {
+
+		if (isLogined == false) {
 			return ResultData.from("F-A", "로그인 후 이용해주세요.");
 		}
-		
+
 		if (Ut.empty(title)) {
 			return ResultData.from("F-1", "title(을)를 입력해주세요.");
 		}
@@ -75,12 +75,29 @@ public class UsrArticleController {
 
 	@RequestMapping("/usr/article/deleteArticle")
 	@ResponseBody
-	public ResultData doDelete(int id) {
+	public ResultData doDelete(HttpSession httpsession, int id) {
+
+		boolean isLogined = false;
+		int loginedMemberId = 0;
+
+		if (httpsession.getAttribute("loginedMemberId") != null) {
+			isLogined = true;
+			loginedMemberId = (int) httpsession.getAttribute("loginedMemberId");
+		}
+
+		if (isLogined == false) {
+			return ResultData.from("F-A", "로그인 후 이용해주세요.");
+		}
 
 		Article article = articleService.getArticle(id);
-
+		
+		
 		if (article == null) {
 			return ResultData.from("F-1", Ut.f("%d번 게시물이 존재하지 않습니다.", id));
+		}
+		
+		if( article.getMemberId() !=  loginedMemberId) {
+			return ResultData.from("F-2", "권한이 없습니다.");
 		}
 
 		articleService.deleteArticle(id);
